gtm/requirements.txt
# Google APIs
google-auth>=2.22.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.1
google-api-python-client>=2.95.0

# HTTP requests
requests>=2.31.0

# Data handling
pandas>=2.0.0
openpyxl>=3.1.0

# Environment
python-dotenv>=1.0.0

# Supabase (for later phases)
supabase>=2.0.0

# Slack (for later phases)
slack-sdk>=3.21.0
slack-bolt>=1.18.0
